Changes command used to wait for pending heals
During the last upgrade, it was noticed that the pending heal count
tended to not decrease until 'gluster v heal info' was run, then it
would suddenly be zero. When investigating this behavior, I was told
the 'heal statistics' family of commands are not supported. The above
problem was seen while looking at pending heals via 'gluster vol heal
statistics heal-count'. Advice was to use 'v heal info summary' instead.
This change implements the above advice.

Signed-off-by: John Strunk <jstrunk@redhat.com>